backend/simple_services/user_schedule/requirements.txt
fastapi==0.110.0
pydantic==2.6.4
python-dotenv==1.0.1
SQLAlchemy==2.0.27
uvicorn==0.29.0
requests==2.31.0
mysql == 0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4


